Progress
I have set my initialState and reducer. I created the action folder and created the index.js inside of that folder. In that index.js component that is where I exported the dispatch actions. Then I went into the AddedFeature.js component and added the onClick function of removeFeature and imported removeFeature from the action folder.