Progress pt 2
I had some help on getting into the right direction. I started to delete the props in the App.js in the return and went the the components that had the props and imported connect and created the mapToStateProps. This was to start taking off a lot of things in App.js and use redux in those other components to display everything again. Now I am going to the rest of the components and go from there.